[IMP] hw_posbox_homepage: user-friendly IoT homepage
Overall improvement of the IoT box homepage.

- All "technical" related features have been moved away from the
main homepage to a dedicated "Technical" page (accessible from
the homepage)
- Icons were added to better represent the homepage information
- (?) with documentation link have been added to more easily find
relevant documentation
- Technical menu navigation to the different technical sub-menus
- Improvement security fix by allowing some sensible URL only if
some conditions are met. E.g: avoid running subprocess meant for
IoT-box on the windows virtual IoT

Visuals are provided in the PR message

task-3893904
