+++ addons/hw_posbox_homepage/views/technical/drivers_settings.jinja2
@@ -1,0 +1,7 @@
+{% extends "technical/_layout.jinja2" %}
+{% block content %}
+    <h2 class="text-center text-green">Drivers Settings</h2>
+
+    {{ iot_table(drivers_table, "No drivers settings found") }}
+
+{% endblock %}
